pyproject.toml
[project]
name = "options-monitor"
version = "0.1.0"
description = "Discord agent for monitoring an options trading bot using Gemini AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.7.1",
    "google-genai>=1.69.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]

[project.scripts]
options-monitor = "options_monitor.main:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/options_monitor"]
